Adjust on initialization of database, implementation of Endpoint to getById and Delete events

+++ src/main/java/com/sportradar/SportCalendar/repository/CardRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardRepository extends JpaRepository<CardEntity, Long> {
+public interface CardRepository extends JpaRepository<CardEntity, Integer> {
 }
